move get_tag_wahoo_xml_path to constants_functions
to be identical with changes from eab03457e1518e59e79d836037c144368e4aae76
