+++ wahoomc/constants_functions.py
@@ -16,6 +16,10 @@ from wahoomc.constants import TOOLING_WIN_DIR
 log = logging.getLogger('main-logger')
 
 
+class TagWahooXmlNotFoundError(Exception):
+    """Raised when the specified tag-wahoo xml file does not exist"""
+
+
 def get_region_of_country(county):
     """
     returns the region / continent of a given country
@@ -156,3 +160,17 @@ def get_tooling_win_path(path_in_tooling_win):
     return path to a tooling in the tooling_win directory and the given path
     """
     return os.path.join(TOOLING_WIN_DIR, *path_in_tooling_win)
+
+
+def get_tag_wahoo_xml_path(tag_wahoo_xml):
+    """
+    return path to tag-wahoo xml file if the file exists
+    """
+
+    path_tag_wahoo_xml = os.path.join(
+        RESOURCES_DIR, "tag_wahoo_adjusted", tag_wahoo_xml)
+
+    if os.path.exists(path_tag_wahoo_xml):
+        return path_tag_wahoo_xml
+
+    raise TagWahooXmlNotFoundError
